added sceleton for sophisticated html report
