+++ ch.hsr.ifs.cdt.metriculator/export-resources/reports/html/dynamic/themes/simple/style.css
