+++ ch.hsr.ifs.cdt.metriculator/export-resources/reports/html/static/themes/simple/style.css
@@ -1,0 +1,108 @@
+html, body {
+            font-family: Verdana;
+            font-size: 14px;
+          	vertical-align: baseline;
+			border: 0 none;
+			outline: 0;
+			padding: 0;
+			margin: 0;
+			text-align: left;
+         }
+         table
+		 {
+		    border-collapse: collapse;
+		    border-spacing: 0;
+		    font-family: "Courier New";
+		 }
+		
+		 img { border: 0; }
+         
+         /*end reset*/
+         
+         table{
+         	margin: 20px;
+         }
+         td, th{
+             padding: 3px;
+             padding-left: 0;
+         }
+         th{
+             font-family: Verdana;
+         }
+         /*preferences*/
+         #preferences td{
+         	font-family: Verdana;
+         }
+         table tr.metric{
+         	font-weight: bold;
+         }
+         table tr.preference td:first-child{
+         	padding-left: 20px;
+         }
+         /*tree*/
+         td.value{
+             font-weight: bold;
+             text-align: right;
+         }
+         th.metric{
+             width: 60px;
+             text-align: right;
+         }
+         td.none{}
+         td.warning{
+         	color: orange;
+         }
+         td.error{
+         	color: red;
+         }
+         tr:hover{
+         	background-color: #eee;
+         }
+         tr.indent-0 td.label{
+             padding-left: 0px;
+         }
+         tr.indent-1 td.label{
+             padding-left: 10px;
+         }
+         tr.indent-2 td.label{
+             padding-left: 20px;
+         }
+         tr.indent-3 td.label{
+             padding-left: 30px;
+         }
+         tr.indent-4 td.label{
+             padding-left: 40px;
+         }
+         tr.indent-5 td.label{
+             padding-left: 50px;
+         }
+         tr.indent-6 td.label{
+             padding-left: 60px;
+         }
+         tr.indent-7 td.label{
+             padding-left: 70px;
+         }
+         tr.indent-8 td.label{
+             padding-left: 80px;
+         }
+         tr.indent-9 td.label{
+             padding-left: 90px;
+         }
+         tr.indent-10 td.label{
+             padding-left: 100px;
+         }
+         tr.indent-11 td.label{
+             padding-left: 110px;
+         }
+         tr.indent-12 td.label{
+             padding-left: 120px;
+         }
+         tr.indent-13 td.label{
+             padding-left: 130px;
+         }
+         tr.indent-14 td.label{
+             padding-left: 140px;
+         }
+         tr.indent-15 td.label{
+             padding-left: 150px;
+         }
