+++ ch.hsr.ifs.cdt.metriculator/src/ch/hsr/ifs/cdt/metriculator/views/reports/TextReportGenerator.java
@@ -1,7 +1,8 @@
-package ch.hsr.ifs.cdt.metriculator.views;
+package ch.hsr.ifs.cdt.metriculator.views.reports;
 
 import java.util.Collection;
 
+import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.Path;
 
 import ch.hsr.ifs.cdt.metriculator.model.AbstractMetric;
@@ -9,9 +10,9 @@ import ch.hsr.ifs.cdt.metriculator.model.converters.ModelToASCIIConverter;
 import ch.hsr.ifs.cdt.metriculator.model.converters.IModelConverter;
 import ch.hsr.ifs.cdt.metriculator.model.nodes.AbstractNode;
 
-public class TextFileExporter extends FileExporter {
+public class TextReportGenerator extends FileReportGenerator {
 
-	public TextFileExporter(Path export_location, AbstractNode root, Collection<AbstractMetric> metrics) {
+	public TextReportGenerator(Path export_location, AbstractNode root, Collection<AbstractMetric> metrics) {
 		super(export_location, root, metrics);
 	}
 
@@ -20,8 +21,9 @@ public class TextFileExporter extends FileExporter {
 		IModelConverter<String> a = new ModelToASCIIConverter();
 		a.convert(root, metrics);
 	
-		// TODO write to file
-		System.out.println(a.getResult());
+		IPath textFilename = export_to_dir.append("metrics").addFileExtension("txt");
+		writeTo(textFilename, a.getResult());
+		open(textFilename);
 	}
 
 }
